Separate Gin IdentityBuilder
