+++ go.work
@@ -5,5 +5,4 @@ use (
 	./middleware/ginz
 	./middleware/gorillaz
 	./middleware/grpcz
-	./middleware/httpz
 )
