+++ middleware/grpcz/go.mod
@@ -6,7 +6,7 @@ replace github.com/aserto-dev/go-aserto => ../../
 
 require (
 	github.com/aserto-dev/errors v0.0.9
-	github.com/aserto-dev/go-aserto v0.32.0
+	github.com/aserto-dev/go-aserto v0.31.5
 	github.com/aserto-dev/go-authorizer v0.20.8
 	github.com/aserto-dev/go-directory v0.31.9
 	github.com/lestrrat-go/jwx/v2 v2.1.1
