--- middleware/httpz/export_test.go
@@ -1,5 +1,0 @@
-package httpz
-
-func InternalHostnameSegment(hostname string, level int) string {
-	return hostnameSegment(hostname, level)
-}
